+++ CHANGELOG.md
@@ -2,4 +2,8 @@
 
 ## 0.5.6
 
-- [Fix] Clear tableview if placeholder is shown
+- **Fix**: Clear tableview if placeholder is shown
+
+## 0.5.7
+
+- **Add**: String extension `localizedComponent(value:stringOne:stringTwo:stringMany:)`